collision update

+++ src/at/ac/tuwien/cg/cgmd/bifth2010/level12/entities/BasicProjectile.java
@@ -3,11 +3,11 @@ package at.ac.tuwien.cg.cgmd.bifth2010.level12.entities;
 
 public class BasicProjectile extends Projectile {
 	
-	public BasicProjectile( float speed, float dmg ){
+	public BasicProjectile( float speed, short dmg ){
 		super( speed, dmg );
 	}
 	
-	public BasicProjectile( float speed, float dmg, float x, float y ){
+	public BasicProjectile( float speed, short dmg, float x, float y ){
 		super( speed, dmg );
 		super.setXY(x, y);
 	}
